backend_python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
Pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
